Fix #8

+++ deploy.sh
@@ -11,8 +11,8 @@ do
   ln -svf $SCRIPT_DIR/$f $HOME/$f
 done
 
-ln -svf $SCRIPT_DIR/bin $HOME/bin
-ln -svf $SCRIPT_DIR/etc $HOME/etc
+[[ ! -e $HOME/bin ]] && ln -svf $SCRIPT_DIR/bin $HOME/bin
+[[ ! -e $HOME/etc ]] && ln -svf $SCRIPT_DIR/etc $HOME/etc
 
 ln -svf $SCRIPT_DIR/dircolors/dircolors.ansi-light $HOME/.dir_colors
